refactor: move some functions to utils and minor changes
