+++ india_compliance/gst_india/doctype/isd_invoice/test_isd_invoice.py
@@ -8,8 +8,8 @@ from erpnext.accounts.utils import get_fiscal_year
 from frappe.tests import IntegrationTestCase
 
 from india_compliance.gst_india.constants import GST_TAX_TYPES
-from india_compliance.gst_india.doctype.isd_invoice.isd_invoice import bulk_create_isd_invoices
 from india_compliance.gst_india.overrides.company import create_company_fixtures
+from india_compliance.gst_india.utils.isd import bulk_create_isd_invoices
 from india_compliance.gst_india.utils.tests import create_purchase_invoice
 from india_compliance.tests.erpnext_test_utils import create_fiscal_year
